refactoring guessnumber/components/foo
