+++ games/guessnumber-core/src/main/java/ch/zhaw/gameconsoleapp/guessnumber/core/annotationinterface/MinNumber.java
@@ -1,4 +1,4 @@
-package ch.zhaw.gameconsoleapp.guessnumber.annotationinterface;
+package ch.zhaw.gameconsoleapp.guessnumber.core.annotationinterface;
 
 import org.springframework.beans.factory.annotation.Qualifier;
