+++ games/guessnumber-core/src/main/java/ch/zhaw/gameconsoleapp/guessnumber/core/components/CoreGameLogicImpl.java
@@ -1,10 +1,10 @@
-package ch.zhaw.gameconsoleapp.guessnumber.service;
+package ch.zhaw.gameconsoleapp.guessnumber.core.components;
 
-import ch.zhaw.gameconsoleapp.guessnumber.annotationinterface.DefaultGuessCount;
+import ch.zhaw.gameconsoleapp.guessnumber.core.annotationinterface.DefaultGuessCount;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
@@ -16,7 +16,7 @@ import javax.annotation.PreDestroy;
  * @author created by Urs Albisser, on 2020-01-07
  * @version 1.0
  */
-@Service
+@Component
 public class CoreGameLogicImpl implements CoreGameLogic {
